requirements/dev.txt
# install for developer environment
-r common.txt
psycopg2-binary==2.9.9
pre-commit==3.6.0
black==23.12.1
ruff==0.1.13
pytest==7.4.4
pytest-django==4.7.0
add-trailing-comma==3.1.0
isort==5.13.2
# development like tests for example
